Fix migration issue by creating new migration.
Both migration #3 and migration #4 in symposion.speakers have not been
applied upstream. Because of this, migration #4 accidentally reverted
the changes made in migration #3.

Because there are currently deployed conference sites with this set of
incorrect migrations, it is necessary to create a new migration to fix
the issue instead of just amending migration #4.