Add help_text to Speaker.photo field.
Add help text to the Photo field in the Speaker model so that users
editing their profile will be warned that image files more than 10 MB
will not work.

Normally, changes like this would be implemented in conf_site, but that
would require substantial modification/duplication of how speakers are
implemented. This, on the other hand, is a one line change.

This patch should not be pushed upstream.